chore: move all pact tests
