+++ packages/template-cli/templates/src/ssr/package.json
@@ -9,10 +9,10 @@
     "lint": "node_modules/.bin/eslint . -f codeframe",
     "prettier": "node_modules/.bin/prettier --ignore-path .gitignore --write \"**/*.+(js|json)\"",
     "validate": "npm run prettier --list-different && npm run lint",
-    "test:pact": "node_modules/.bin/jest --c jest-pact.config.json --rootDir __tests__ --runInBand",
-    "test:pact-publish": "node_modules/.bin/tsc ./__tests__/pact/pactPublish.ts && node ./__tests__/pact/pactPublish.js",
-    "test:pact-start-stub": "node_modules/.bin/tsc ./__tests__/pact/packStubServer.ts && node __tests__/pact/packStubServer.js",
-    "test:pact-can-i-deploy": "node_modules/.bin/tsc ./__tests__/pact/pactCanIDeploy.ts && node __tests__/pact/pactCanIDeploy.js",
+    "test:pact": "node_modules/.bin/jest --c ./__tests__/pact/jest-pact.config.json --rootDir __tests__/pact --runInBand",
+    "test:pact-publish": "node_modules/.bin/tsc ./__tests__/pact/utils/pactPublish.ts && node ./__tests__/pact/utils/pactPublish.js",
+    "test:pact-start-stub": "node_modules/.bin/tsc ./__tests__/pact/utils/pactStubServer.ts && node __tests__/pact/utils/pactStubServer.js",
+    "test:pact-can-i-deploy": "node_modules/.bin/tsc ./__tests__/pact/utils/pactCanIDeploy.ts && node __tests__/pact/utils/pactCanIDeploy.js",
     "test:pact-can-i-deploy-ci": "node_modules/.bin/pact-broker can-i-deploy --pacticipant $PACT_CONSUMER -l --broker-base-url $PACT_BROKER --broker-token $PACT_BEARER_TOKEN",
     "test:cypress:watch": "node_modules/.bin/tsc --watch --project __tests__/cypress/tsconfig.cypress.json",
     "test:cypress:compile": "node_modules/.bin/tsc --project __tests__/cypress/tsconfig.cypress.json",
